// src/components/pages/MaintenanceData/MaintenanceDebtor/Documentation/components/ModalDetailDocument/ModalDetailDocument.types.ts
export type ModalDetailDocumentProps = {
  data: {
    bucketMasterId: string;
    bucketProcessId: string;
    documentGroup: string;
    documentType: string;
    documentName: string;
    documentNumber: string;
    documentDate: string;
    uploadedBy: string;
    uploadedDate: string;
    creatorName: string;
    divisionLabel: string;
    document: any;
    documentTypeLabel: string;
    fileName: string;
    documentGroupLabel: string;
  };
}